fix(core): simplify DB initialization and move worktree logic
- Remove metadata table — schema version is encoded in the DB filename,
  making the runtime version check redundant
- Simplify initialize_db to use db_existed_before_open instead of
  querying sqlite_master
- Move get_main_worktree_root into its own worktree module (no wasm gate
  needed since it's just git commands)
